+++ mapper/lib/src/model/index.dart
@@ -5,16 +5,12 @@ import 'type_info.dart';
 export 'adapters.dart';
 export 'annotations.dart';
 export 'converters.dart';
+export 'enum.dart';
 export 'name_casing.dart';
 export 'type_info.dart';
 export 'value_decorators.dart';
 
 enum ConversionDirection { fromJson, toJson }
-
-/// Virtual class
-/// Used as a generic reference to all Enum based types
-/// enum ABC {A, B, C}, etc.
-abstract class Enum {}
 
 const defaultDeserializationOptions = DeserializationOptions();